requirements.txt
Pillow>=10.0
transformers>=4.48.0
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.9
safetensors>=0.4.5
scipy>=1.11.0
